editor-page moved to challenges-module
