+++ frontend/src/app/modules/challenges/online-editor/online-editor-page/online-editor-page.component.html
@@ -34,7 +34,7 @@
                             [isMultiSelection]="false"
                             [items]="languages"
                             [selectedItem]="selectedLanguage"
-                            (SelectedItemsChanged)="onSelectedLanguageChanged($event)">
+                            (selectedItemsChanged)="onSelectedLanguageChanged($event)">
                         </app-dropdown-select>
                         <app-dropdown-select
                             [isMultiSelection]="false"
