+++ frontend/src/app/modules/challenges/online-editor/online-editor-page/online-editor-page.component.ts
